+++ CHANGELOG.md
@@ -23,6 +23,7 @@ This changelog only contains the changes that are unreleased. For changes for in
 - Auto select latest file when updating mods
 - FTB pack links not working [#681]
 - Update LaunchServer.bat to work with double quotes in javapath and print more info [#682]
+- Issue with Arm based Mac not using Java runtimes [#684]
 
 ### Misc
 - Implement view model for InstancesTab [#675]